refactor: split GoesData.tsx into separate component files
Extract 8 components from the monolithic 1631-line GoesData.tsx:
- types.ts: shared interfaces (Product, GoesFrame, AnimationType, etc.)
- utils.ts: formatBytes helper
- FetchTab.tsx: satellite data fetching with gap detection
- BrowseTab.tsx: frame browser with filters, grid/list views
- FramePreviewModal.tsx: image preview with crop tool
- AddToCollectionModal.tsx: add frames to collections
- TagModal.tsx: tag management for frames
- AnimationStudioTab.tsx: animation creation and history
- CollectionsTab.tsx: collection CRUD
- StatsTab.tsx: storage and frame statistics

GoesData.tsx reduced from 1631 to ~105 lines (imports + tab routing).
Tab bar now grouped into Data | Tools | Manage categories with visual dividers.

Pure refactor - no functionality changes.
